packages/jb-translator/pyproject.toml
[tool.poetry]
name = "jb-translator"
version = "0.1.0"
description = ""
authors = ["OpenNyAI Team <opennyai@googlegroups.com>"]
readme = "README.md"
packages = [{include = "jugalbandi/translator"}]

[tool.poetry.dependencies]
python = ">=3.10, <4.0.0"
google-cloud-translate = "^3.11.1"
httpx = "^0.24.1"
jb-core = {path = "../jb-core", develop = true}
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"
poethepoet = "^0.20.0"
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poe.tasks.lint]
shell = """
black jugalbandi tests
flake8 jugalbandi tests
mypy jugalbandi tests
"""
interpreter = "bash"
help = "format, lint, typecheck"


[tool.poe.tasks.test]
cmd = "python -m pytest -vv -o log_cli=1 -o log_cli_level=INFO -W 'ignore::DeprecationWarning' $FILE"
args = [{name="FILE", default="tests", positional=true}]
help = "run tests using pytest"
